AV-645: moved organization approval stuff to own module
