+++ modules/ckanext-organizationapproval/.coveragerc
@@ -1,0 +1,5 @@
+[report]
+omit =
+    */site-packages/*
+    */python?.?/*
+    ckan/*
