+++ modules/ckanext-organizationapproval/.travis.yml
@@ -1,0 +1,13 @@
+language: python
+sudo: required
+python:
+    - "2.7"
+services:
+    - postgresql
+    - redis-server
+install:
+    - bash bin/travis-build.bash
+    - pip install coveralls
+script: sh bin/travis-run.sh
+after_success:
+    - coveralls
