+++ modules/ckanext-organizationapproval/MANIFEST.in
@@ -1,0 +1,4 @@
+include README.rst
+include LICENSE
+include requirements.txt
+recursive-include ckanext/organizationapproval *.html *.json *.js *.less *.css *.mo
