+++ modules/ckanext-organizationapproval/ckanext/organizationapproval/templates/admin/manage_organizations.html
@@ -36,7 +36,6 @@
           <span>
             {{ _('To make an existing user a Showcase Admin, search for their username below.') }}
           </span>
-
           <div class="controls">
             <input id="username" type="text" name="username" placeholder="{{_('Username')}}"
             value="" class="control-medium" data-module="autocomplete"
@@ -44,7 +43,6 @@
           </div>
         </div>
       </div>
-
       <div class="form-actions">
         <button class="btn btn-primary" type="submit" name="submit" >
           {{ _('Add User') }}
