+++ modules/ckanext-organizationapproval/ckanext/organizationapproval/tests/test_plugin.py
@@ -1,0 +1,5 @@
+"""Tests for plugin.py."""
+import ckanext.organizationapproval.plugin as plugin
+
+def test_plugin():
+    pass
