+++ modules/ckanext-organizationapproval/setup.cfg
@@ -1,0 +1,21 @@
+[extract_messages]
+keywords = translate isPlural
+add_comments = TRANSLATORS:
+output_file = ckanext/organizationapproval/i18n/ckanext-organizationapproval.pot
+width = 80
+
+[init_catalog]
+domain = ckanext-organizationapproval
+input_file = ckanext/organizationapproval/i18n/ckanext-organizationapproval.pot
+output_dir = ckanext/organizationapproval/i18n
+
+[update_catalog]
+domain = ckanext-organizationapproval
+input_file = ckanext/organizationapproval/i18n/ckanext-organizationapproval.pot
+output_dir = ckanext/organizationapproval/i18n
+previous = true
+
+[compile_catalog]
+domain = ckanext-organizationapproval
+directory = ckanext/organizationapproval/i18n
+statistics = true
